% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelesToMutations.R
\name{allelesToMutations}
\alias{allelesToMutations}
\title{allelesToMutations}
\usage{
allelesToMutations(alpath)
}
\arguments{
\item{alpath}{path to one Alleles_frequency_table, created by CRISPResso2}
}
\description{
allelesToMutations
}
