% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelesToMutations.R
\name{alleleToMutation}
\alias{alleleToMutation}
\title{alleleToMutation}
\usage{
alleleToMutation(ref, ali, nreads, rowi)
}
\arguments{
\item{ref}{one aligned reference, split}

\item{ali}{one aligned sequence, split}

\item{rowi}{}
}
\description{
alleleToMutation
}
