% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callMutations.R
\name{callMutations}
\alias{callMutations}
\title{callMutations}
\usage{
callMutations(
  copath,
  metapath,
  minnreads = NA,
  controltb = NA,
  callSubs = TRUE,
  exportpath
)
}
\arguments{
\item{copath}{folder that contains the CRISPResso results directories}

\item{exportpath}{}
}
\description{
callMutations
}
