% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyReads.R
\name{classifyReads}
\alias{classifyReads}
\title{classifyReads}
\usage{
classifyReads(
  mut,
  mode = "precise",
  scaffDetect = FALSE,
  whichScaff = "std",
  scaffWin = NA,
  unwantedSubs = FALSE,
  nreadsSample = NA,
  exportpath
)
}
\arguments{
\item{exportpath}{}
}
\description{
classifyReads
}
