% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectTemplatedIns.R
\name{detectTemplatedIns}
\alias{detectTemplatedIns}
\title{detectTemplatedIns}
\usage{
detectTemplatedIns(
  mut,
  allowStretchMatches = 5,
  extendNewlySeq = 3,
  searchWindowStarts = 20,
  minLCSbp = 5
)
}
\arguments{
\item{minLCSbp}{}
}
\description{
detectTemplatedIns
}
