% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectTemplatedIns.R
\name{detectTemplatedIns_one}
\alias{detectTemplatedIns_one}
\title{detectTemplatedIns_one}
\usage{
detectTemplatedIns_one(
  alrow,
  minLCSbp,
  allowStretchMatches,
  extendNewlySeq,
  searchWindowStarts
)
}
\arguments{
\item{searchWindowStarts}{}
}
\description{
detectTemplatedIns_one
}
