% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMutations.R
\name{filterMutations}
\alias{filterMutations}
\title{filterMutations}
\usage{
filterMutations(
  muttb,
  minnreads = NA,
  cutpos = NA,
  cutdist = NA,
  rhapos = NA,
  rhadist = NA,
  controltb = NA,
  callSubs = TRUE
)
}
\arguments{
\item{callSubs}{}
}
\description{
filterMutations
}
