% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggDelDot.R
\name{ggDelDot}
\alias{ggDelDot}
\title{ggDelDot}
\usage{
ggDelDot(
  mut,
  colMHdel = "#c77f93",
  minMHlen = 1,
  xtextOrNo = TRUE,
  xnameOrNo = TRUE,
  ynameOrNo = TRUE,
  legendOrNo = TRUE,
  marginTB = NA,
  exportPath = NA,
  width = 60,
  height = 50
)
}
\arguments{
\item{height}{}
}
\description{
ggDelDot
}
