% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggIGVdel.R
\name{ggIGVdel}
\alias{ggIGVdel}
\title{ggIGVdel}
\usage{
ggIGVdel(
  mut,
  rangeStart = NA,
  rangeStop = NA,
  bpSides = NA,
  titleOrNo = TRUE,
  orderdels = "start",
  ligPos = FALSE,
  forceligPos = NA,
  avgLigPos = FALSE,
  addBottomMargin = 0,
  exportPath,
  width,
  height
)
}
\arguments{
\item{height}{}
}
\description{
ggIGVdel
}
