% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMHdel.R
\name{ggMHdel}
\alias{ggMHdel}
\title{ggMHdel}
\usage{
ggMHdel(
  mut,
  min_del_nreads = 50,
  colourLight = "#f1b9c7",
  colourManual = NA,
  splitby = "grp",
  onlygrp = NA,
  grporder = NA,
  legendOrNo = TRUE,
  titleOrNo = TRUE,
  xtextOrNo = TRUE,
  ytextOrNo = TRUE,
  ynameOrNo = TRUE,
  panelSpacing = NA,
  exportpath,
  width = 110,
  height = 65
)
}
\arguments{
\item{height}{}
}
\description{
ggMHdel
}
