% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggStack.R
\name{ggStack}
\alias{ggStack}
\title{ggStack}
\usage{
ggStack(
  rlab,
  splitby = "locus",
  onlyrundate = "all",
  onlylocus = "all",
  onlygrp = "all",
  onlycat = "all",
  grporder = NA,
  locusorder = NA,
  catorder = NA,
  catcols = NA,
  mincov = NA,
  ymax = 1,
  titleOrNo = TRUE,
  xtextOrNo = TRUE,
  ytextOrNo = TRUE,
  ynameOrNo = TRUE,
  legendOrNo = TRUE,
  splcovOrNo = FALSE,
  labelSamples = NA,
  xgrid = FALSE,
  titleSize = 9,
  ytitleSize = 9,
  ytextSize = 7,
  panelSpacing = NA,
  exportOrNo = TRUE,
  logOrNo = FALSE,
  width = 159.1,
  height = 82.03,
  exportpath
)
}
\arguments{
\item{exportpath}{}
}
\description{
ggStack
}
