% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggTemplIns.R
\name{ggTemplIns}
\alias{ggTemplIns}
\title{ggTemplIns}
\usage{
ggTemplIns(
  mut,
  min_ins_nreads = 50,
  colourLow = "#e4eee0",
  colourHigh = "#304528",
  grporder = NA,
  legendOrNo = TRUE,
  titleOrNo = TRUE,
  xtextOrNo = TRUE,
  ytextOrNo = TRUE,
  ynameOrNo = TRUE,
  exportpath,
  width = 110,
  height = 65
)
}
\arguments{
\item{height}{}
}
\description{
ggTemplIns
}
