% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectTemplatedIns.R
\name{newlySynthesised}
\alias{newlySynthesised}
\title{newlySynthesised}
\usage{
newlySynthesised(alrow, allowStretchMatches, extendNewlySeq)
}
\arguments{
\item{extendNewlySeq}{}
}
\description{
newlySynthesised
}
