% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preciseClassify_one.R
\name{preciseClassify_one}
\alias{preciseClassify_one}
\title{preciseClassify_one}
\usage{
preciseClassify_one(
  mut,
  scaffDetect,
  pestrand,
  whichScaff,
  scaffWin,
  unwantedSubs
)
}
\arguments{
\item{unwantedSubs}{}
}
\description{
preciseClassify_one
}
