% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelesToMutations.R
\name{recordDel}
\alias{recordDel}
\title{recordDel}
\usage{
recordDel(ref, ali)
}
\arguments{
\item{ref}{reference sequence (aligned), split}

\item{ali}{aligned sequence, split}
}
\description{
recordDel
}
