% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelesToMutations.R
\name{recordIns}
\alias{recordIns}
\title{recordIns}
\usage{
recordIns(ref, ali)
}
\arguments{
\item{ref}{reference sequence (aligned), split}

\item{ali}{aligned sequence, split}
}
\description{
recordIns
}
