% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelesToMutations.R
\name{recordRef}
\alias{recordRef}
\title{recordRef}
\usage{
recordRef(ref, ali)
}
\arguments{
\item{ref}{reference sequence (aligned), split}

\item{ali}{aligned sequence, split}
}
\description{
recordRef
}
