% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectTemplatedIns.R
\name{searchLCS}
\alias{searchLCS}
\title{searchLCS}
\usage{
searchLCS(
  alrow,
  newlyseq,
  newlyseqStart,
  newlyseqStop,
  oref,
  minLCSbp,
  searchWindowStarts,
  cutpos = NA
)
}
\arguments{
\item{cutpos}{}
}
\description{
searchLCS
}
