% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateDel.R
\name{simulateDel}
\alias{simulateDel}
\title{simulateDel}
\usage{
simulateDel(
  mut,
  nreads = 1000,
  cutDelbp = 3,
  awayfromCut = 4,
  fit_meanlog = 1.955957,
  fit_sdlog = 0.8970051
)
}
\arguments{
\item{fit_sdlog}{}
}
\description{
simulateDel
}
