% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateDel.R
\name{simulateDel_one}
\alias{simulateDel_one}
\title{simulateDel_one}
\usage{
simulateDel_one(mut, fit_meanlog, fit_sdlog, cutpos, cutDelbp, awayfromCut)
}
\arguments{
\item{awayfromCut}{}
}
\description{
simulateDel_one
}
