% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateIns.R
\name{simulateIns}
\alias{simulateIns}
\title{simulateIns}
\usage{
simulateIns(
  mutdti,
  nreads = 1000,
  extendNewlySeq = 0,
  searchWindowStarts = 20,
  minLCSbp = 5,
  fit_exprate = 0.06643421
)
}
\arguments{
\item{fit_exprate}{}
}
\description{
simulateIns
}
