% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateIns.R
\name{simulateIns_one}
\alias{simulateIns_one}
\title{simulateIns_one}
\usage{
simulateIns_one(
  extendNewlySeq,
  searchWindowStarts,
  minLCSbp,
  cutpos,
  oref,
  fit_exprate
)
}
\arguments{
\item{fit_exprate}{}
}
\description{
simulateIns_one
}
